moved memory allocation only to necessary spot
ASAN had detected an issue previously with line 565. This was allocating a vector of indices for the short
genotype. It looks like I do not free this memory and that's why it's complaining, but it might be a deeper issue. Since this information is only needed for the genome addition model, I have moved it into that controll structure.