fix NULL population bug in plot_poppr_msn()
There was a bug where data sets with no population
defined could not be plotted with
`plot_poppr_msn()`. This was due to the fact that
it would run through `palette_parser()` and
`update_poppr_graph()`, which replaces the
population colors given a palette. Unfortunately,
it turns out when you have a value of NA for a
population, `palette_parser()` will still name the
resulting vector with NA, and the data frame will
freak out because it's getting a row name that's
NA (because it cares about such silliness). In any
case, this fixes that bug.