update to fix #109
It turns out that that error was actually coming
from inside the function. It was just not very
informative to the user. I've changed this and
explicitly tested it.

I've also come across and fixed another bug where
the incoming distance function was never actually
being evaluated in the first loop to ensure no one
used missingno with diss.dist. This has also been
fixed in here.