rename C function genotype_curve to .._internal
This is to further address #124. Apparently, without
adding .registration = TRUE, the NOTE will go away,
but nothing will actually be registered.

Once I added .registration = TRUE (in 8cfc29e4),
genotype_curve was flagged with this warning
(in https://travis-ci.org/grunwaldlab/poppr/jobs/214713382#L837-L843):

Warning: failed to assign RegisteredNativeSymbol for genotype_curve to genotype_curve since genotype_curve is already defined in the ‘poppr’ namespace
A namespace must be able to be loaded with just the base namespace
loaded: otherwise if the namespace gets loaded by a saved object, the
session will be unable to start.
Probably some imports need to be declared in the NAMESPACE file.