Add more specificity to detect adegenet renaming
An issue was brought up by 刁永朝 where he was missing six populations
when he attempted to do aboot.

It turned out that he had a duplicated sample name in his data and
identified some of his samples with numbers that were within the range
of samples he had. Adegenet will rename samples with numbers and I
foolishly checked if any of the sample names were present in the genind
object, not thinking that numeric names will match up.

This fix changes the test to see if the intersection between the genind
object and the original names provides the names in the genind object.
If they are not equal, then the integer names were used.