Change branch name; rename option
Reference: https://www.hanselman.com/blog/EasilyRenameYourGitDefaultBranchFromMasterToMain.aspx

I've changed the default branch from "master" to "main" to remove a
metaphor of oppression (if only in a symbolic manner).

If you have a local copy of poppr on your system, you can update the
branch by using the following commands:

 git checkout master
 git branch -m master main
 git fetch
 git branch --unset-upstream
 git branch -u origin/main
 git symbolic-ref refs/remotes/origin/HEAD refs/remotes/origin/main

I've also deprecated "blacklist" in favor of "exclude" because that is a
clearer descriptor of the option.