update expectation for test to fix #241
It's worth noting that this test was _not incorrect_ numerically because
it was only considering the output from the distance matrix, which was a
lower triangle. Because the values that are recycled end up being thrown
away in the distance object, it was okay for us to construct one without
needing to supply the other side of the matrix.

This fixes the issue by adding a helper function to fill in the correct
values of the matrix and then convert to a distance to avoid the error
thrown by R when not supplying a vector with the correct number of
values for the matrix